# shared/openapi.yaml
openapi: 3.0.3
info:
  title: Kongre Beacon API
  description: Kongre Beacon Sistemi backend API sozlesmesi.
  version: 0.1.0
servers:
  - url: http://localhost:3001
    description: Yerel gelistirme
paths:
...
  /attendance/summary:
    get:
      summary: Kongre icin canli katilim/doluluk ozeti (panel KPI'lari)
      operationId: getAttendanceSummary
      tags:
        - Attendance
      parameters:
        - name: congressId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ozet veri.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceSummary'
  /attendance/hall-visits:
    get:
      summary: Katilimci salon ziyaret gecmisi (sayfalanmis)
      operationId: listHallVisits
      tags:
        - Attendance
      parameters:
        - name: congressId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: hallId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: userId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: isOpen
          in: query
          required: false
          schema:
            type: boolean
        - name: search
          in: query
          required: false
          description: Katilimci ad/soyad icinde arama (kismi eslesme).
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Sayfalanmis ziyaret listesi.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HallVisitPage'
  /attendance/occupancy-series:
    get:
      summary: Zaman-dilimli salon doluluk serisi (yogunluk grafigi icin)
      description: >-
        HallVisit araliklarindan hesaplanir; her bucket icin salon basina
        o an icerideki tahmini kisi sayisini doner. `from`/`to` verilmezse
        varsayilan bugun 00:00 - simdi araligidir.
      operationId: getOccupancySeries
      tags:
        - Attendance
      parameters:
        - name: congressId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: hallId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: bucketMinutes
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1440
            default: 15
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Zaman serisi verisi.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OccupancySeries'
...
  /congresses:
    post:
      summary: Yeni kongre olustur
      operationId: createCongress
      tags:
        - Congress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CongressCreate'
      responses:
        '201':
          description: Kongre olusturuldu.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Congress'
        '400':
          description: Gecersiz istek govdesi.
    get:
      summary: Kongreleri listele
      operationId: listCongresses
      tags:
        - Congress
      responses:
        '200':
          description: Kongre listesi.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Congress'
  /congresses/{id}:
    get:
      summary: Tek bir kongreyi getir
      operationId: getCongress
      tags:
        - Congress
      parameters:
        - $ref: '#/components/parameters/CongressId'
      responses:
        '200':
          description: Kongre bulundu.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Congress'
        '404':
          description: Kongre bulunamadi.
    patch:
      summary: Kongreyi guncelle
      operationId: updateCongress
      tags:
        - Congress
      parameters:
        - $ref: '#/components/parameters/CongressId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CongressUpdate'
      responses:
        '200':
          description: Kongre guncellendi.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Congress'
        '404':
          description: Kongre bulunamadi.
    delete:
      summary: Kongreyi sil
      operationId: deleteCongress
      tags:
        - Congress
      parameters:
        - $ref: '#/components/parameters/CongressId'
      responses:
        '204':
          description: Kongre silindi.
        '404':
          description: Kongre bulunamadi.
...
  /halls/{id}:
    get:
      summary: Tek bir salonu getir
      operationId: getHall
      tags:
        - Hall
      parameters:
        - $ref: '#/components/parameters/HallId'
      responses:
        '200':
          description: Salon bulundu.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hall'
        '404':
          description: Salon bulunamadi.
    patch:
      summary: Salonu guncelle (isim, rssiThreshold)
      operationId: updateHall
      tags:
        - Hall
      parameters:
        - $ref: '#/components/parameters/HallId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HallUpdate'
      responses:
        '200':
          description: Salon guncellendi.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hall'
        '404':
          description: Salon bulunamadi.
    delete:
      summary: Salonu sil
      operationId: deleteHall
      tags:
        - Hall
      parameters:
        - $ref: '#/components/parameters/HallId'
      responses:
        '204':
          description: Salon silindi.
        '404':
          description: Salon bulunamadi.
  /beacons:
    post:
      summary: Yeni beacon olustur
      operationId: createBeacon
      tags:
        - Beacon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeaconCreate'
      responses:
        '201':
          description: Beacon olusturuldu.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Beacon'
        '400':
          description: Gecersiz istek govdesi.
        '404':
          description: congressId ile eslesen kongre bulunamadi.
        '409':
          description: Ayni kongrede ayni (uuid, major, minor) kombinasyonu zaten var.
    get:
      summary: Beacon'lari listele
      operationId: listBeacons
      tags:
        - Beacon
      parameters:
        - name: congressId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Beacon listesi.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Beacon'
  /beacons/{id}:
    get:
      summary: Tek bir beacon'i getir
      operationId: getBeacon
      tags:
        - Beacon
      parameters:
        - $ref: '#/components/parameters/BeaconId'
      responses:
        '200':
          description: Beacon bulundu.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Beacon'
        '404':
          description: Beacon bulunamadi.
    patch:
      summary: Beacon'i guncelle (congressId haric)
      operationId: updateBeacon
      tags:
        - Beacon
      parameters:
        - $ref: '#/components/parameters/BeaconId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeaconUpdate'
      responses:
        '200':
          description: Beacon guncellendi.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Beacon'
        '404':
          description: Beacon bulunamadi.
        '409':
          description: Ayni kongrede ayni (uuid, major, minor) kombinasyonu zaten var.
    delete:
      summary: Beacon'i sil
      operationId: deleteBeacon
      tags:
        - Beacon
      parameters:
        - $ref: '#/components/parameters/BeaconId'
      responses:
        '204':
          description: Beacon silindi.
        '404':
          description: Beacon bulunamadi.
  /halls/{hallId}/beacons:
    get:
      summary: Salona aktif olarak atanmis beacon'lari listele
      operationId: listActiveHallBeacons
      tags:
        - HallBeacon
      parameters:
        - $ref: '#/components/parameters/HallIdPath'
      responses:
        '200':
          description: Aktif atama listesi.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HallBeacon'
        '404':
          description: Salon bulunamadi.
  /halls/{hallId}/beacons/{beaconId}:
    post:
      summary: Beacon'i salona ata
      description: >-
        Beacon'in zaten baska bir salonda aktif bir atamasi varsa 409 doner;
        once o atama DELETE ile kaldirilmalidir.
      operationId: assignHallBeacon
      tags:
        - HallBeacon
      parameters:
        - $ref: '#/components/parameters/HallIdPath'
        - $ref: '#/components/parameters/BeaconIdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HallBeaconAssign'
      responses:
        '201':
          description: Atama olusturuldu.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HallBeacon'
        '404':
          description: Salon veya beacon bulunamadi.
        '409':
          description: Beacon zaten baska bir salona aktif olarak atanmis.
    delete:
      summary: Beacon'in bu salondaki aktif atamasini kaldir
      operationId: unassignHallBeacon
      tags:
        - HallBeacon
      parameters:
        - $ref: '#/components/parameters/HallIdPath'
        - $ref: '#/components/parameters/BeaconIdPath'
      responses:
        '200':
          description: Atama kaldirildi (isActive=false, removedAt set edildi).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HallBeacon'
        '404':
          description: Bu salon icin aktif bir atama bulunamadi.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CongressId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    HallId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    BeaconId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    HallIdPath:
      name: hallId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    BeaconIdPath:
      name: beaconId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    HallOccupancy:
      type: object
      required:
        - hallId
        - hallName
        - count
      properties:
        hallId:
          type: string
          format: uuid
        hallName:
          type: string
        count:
          type: integer
...
    AttendanceSummary:
      type: object
      required:
        - activeHalls
        - currentlyInsideTotal
        - hallOccupancy
        - participantsSeenToday
        - lastObservationAt
        - durationStats
      properties:
        activeHalls:
          type: integer
        currentlyInsideTotal:
          type: integer
        hallOccupancy:
          type: array
          items:
            $ref: '#/components/schemas/HallOccupancy'
        participantsSeenToday:
          type: integer
        lastObservationAt:
          type: string
          format: date-time
          nullable: true
        durationStats:
          type: array
          items:
            type: object
            required:
              - hallId
              - hallName
              - visitCount
              - averageMinutes
              - medianMinutes
            properties:
              hallId:
                type: string
                format: uuid
              hallName:
                type: string
              visitCount:
                type: integer
              averageMinutes:
                type: number
                nullable: true
              medianMinutes:
                type: number
                nullable: true
    HallVisitSummary:
      type: object
      required:
        - id
        - userFirstName
        - userLastName
        - hallId
        - hallName
        - startedAt
        - isOpen
        - algorithmVersion
      properties:
        id:
          type: string
          format: uuid
        userFirstName:
          type: string
        userLastName:
          type: string
        hallId:
          type: string
          format: uuid
        hallName:
          type: string
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        isOpen:
          type: boolean
        confidenceLevel:
          type: string
          nullable: true
        algorithmVersion:
          type: string
    HallVisitPage:
      type: object
      required:
        - items
        - total
        - page
        - pageSize
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/HallVisitSummary'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    OccupancySeriesPoint:
      type: object
      required:
        - bucketStart
        - values
      properties:
        bucketStart:
          type: string
          format: date-time
        values:
          type: object
          description: hallId -> o bucket'taki tahmini kisi sayisi.
          additionalProperties:
            type: integer
    OccupancySeries:
      type: object
      required:
        - from
        - to
        - bucketMinutes
        - halls
        - points
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        bucketMinutes:
          type: integer
        halls:
          type: array
          items:
            type: object
            required:
              - hallId
              - hallName
            properties:
              hallId:
                type: string
                format: uuid
              hallName:
                type: string
        points:
          type: array
          items:
            $ref: '#/components/schemas/OccupancySeriesPoint'
...
    Congress:
      type: object
      required:
        - id
        - name
        - code
        - accessCode
        - beaconUuid
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
          description: Kisa, insan-okunabilir kongre kodu (pilot login icin kullanilir).
        accessCode:
          type: string
          description: Pilot ortak erisim kodu (gercek kongrede kayit sistemiyle degisecek).
        beaconUuid:
          type: string
          format: uuid
          description: Kongreye ait ortak iBeacon UUID'i.
        startDate:
          type: string
          format: date-time
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CongressCreate:
      type: object
      required:
        - name
        - code
        - accessCode
        - beaconUuid
      properties:
        name:
          type: string
          minLength: 2
        code:
          type: string
          minLength: 2
        accessCode:
          type: string
          minLength: 4
        beaconUuid:
          type: string
          format: uuid
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
    CongressUpdate:
      type: object
      description: CongressCreate ile ayni alanlar, hepsi opsiyonel (PATCH).
      properties:
        name:
          type: string
          minLength: 2
        code:
          type: string
          minLength: 2
        accessCode:
          type: string
          minLength: 4
        beaconUuid:
          type: string
          format: uuid
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
    Hall:
      type: object
      required:
        - id
        - congressId
        - name
        - rssiThreshold
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        congressId:
          type: string
          format: uuid
        name:
          type: string
        rssiThreshold:
          type: integer
          description: Salon icinde kabul edilen minimum RSSI (dBm). Varsayilan -70.
          example: -70
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    HallUpdate:
      type: object
      description: HallCreate ile ayni alanlar (congressId haric), hepsi opsiyonel (PATCH).
      properties:
        name:
          type: string
          minLength: 2
        rssiThreshold:
          type: integer
          minimum: -100
          maximum: 0
    Beacon:
      type: object
      required:
        - id
        - congressId
        - uuid
        - major
        - minor
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        congressId:
          type: string
          format: uuid
        uuid:
          type: string
          format: uuid
          description: Fiziksel beacon cihazinin iBeacon UUID'i.
        major:
          type: integer
          minimum: 0
          maximum: 65535
        minor:
          type: integer
          minimum: 0
          maximum: 65535
        label:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BeaconCreate:
      type: object
      required:
        - congressId
        - uuid
        - major
        - minor
      properties:
        congressId:
          type: string
          format: uuid
        uuid:
          type: string
          format: uuid
        major:
          type: integer
          minimum: 0
          maximum: 65535
        minor:
          type: integer
          minimum: 0
          maximum: 65535
        label:
          type: string
    BeaconUpdate:
      type: object
      description: BeaconCreate ile ayni alanlar (congressId haric), hepsi opsiyonel (PATCH).
      properties:
        uuid:
          type: string
          format: uuid
        major:
          type: integer
          minimum: 0
          maximum: 65535
        minor:
          type: integer
          minimum: 0
          maximum: 65535
        label:
          type: string
    HallBeacon:
      type: object
      description: Bir beacon'in belirli bir salona atanmasini temsil eder.
      required:
        - id
        - hallId
        - beaconId
        - isActive
        - installedAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        hallId:
          type: string
          format: uuid
        beaconId:
          type: string
          format: uuid
        isActive:
          type: boolean
        installedAt:
          type: string
          format: date-time
        removedAt:
          type: string
          format: date-time
          nullable: true
        rssiThreshold:
          type: integer
          nullable: true
          description: Salonun varsayilan esigini bu atama icin override eder.
        calibrationNote:
          type: string
          nullable: true
        placementNote:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        beacon:
          $ref: '#/components/schemas/Beacon'
    HallBeaconAssign:
      type: object
      properties:
        rssiThreshold:
          type: integer
          minimum: -100
          maximum: 0
        calibrationNote:
          type: string
        placementNote:
          type: string
...